Refactor decode
All helper methods now take a { any, binary } as their first argument.
They pass this around, keeping the memoized value in the first position
and the remaining JSON to parse in the second.

Converted some multiple definitions to case statements since Elixir
cannot match inside the string in a tuple.